requirements.txt
discord.py
apscheduler
pyjokes
python-dotenv